[SwipeableDrawer] Extract swiping feature from Drawer
